Separate definitions of Street methods
